refactor: create mosaic fastify plugins
