+++ packages/cli/package.json
@@ -53,6 +53,7 @@
     "cors": "^2.8.5",
     "deepmerge": "^4.2.2",
     "fastify": "^4.23.2",
+    "fastify-plugin": "^4.5.1",
     "fs-extra": "^10.1.0",
     "globby": "^13.1.3",
     "mkdirp": "^1.0.4"
